fix: Use 3/4 sampling with B210
When running the DU with a B210, the error `Error: unknown sampling rate
61440000.000000` appeared in the logs. After some quick search, it turns
out that this series of radio needs to be run at 3/4 sampling, with the
`-E` command-line switch.

This commit hard codes this switch in the charm, as we are only
currently testing with the B210 radio. This will need to be exposed as
a configuration item in the future, but the way to do so will need
proper design.